+++ app/routes/sketch+/artboards+/$slug_+/components/panel/artboard/panel-content-artboard-background-color.tsx
@@ -4,8 +4,8 @@ import {
 	PanelRowContainer,
 	PanelTitle,
 } from '#app/components/shared'
-import { type PickedArtboardType } from '../queries'
-import { PanelFormArtboardEditBackgroundColor } from './panel-form-artboard-background-color'
+import { type PickedArtboardType } from '../../../queries'
+import { PanelFormArtboardEditBackgroundColor } from '../../forms/artboard/panel-form-artboard-edit-background-color'
 
 export const PanelContentArtboardBackgroundColor = ({
 	artboard,
